🎨 refact: separate PopupViewControlle from MapViewController
